fix: CodeQL highs — snprintf truncation clamps + loop-counter width
Three genuine correctness fixes flagged by CodeQL on this PR:

- benchmark_decomp.cpp: every 'n += snprintf(b+n, cap-n, ...)'
  accumulation could push n past cap on truncation and underflow
  the next size argument. All sites now go through snappend(),
  which clamps (vsnprintf-based, same format args).
- benchmarks/dom/bench_text_borrowed.c: the original instance of
  the same pattern (cpp/overflowing-snprintf, alert #30) — clamped
  with the same discipline.
- element.c taurus_element_children/indexed walk: the loop counter
  was uint8_t against a size_t index — wraps for index > 255 and
  the early-exit never fires as intended (comparison-with-wider-
  type, alert #7). Now size_t.

567/567 tests.

+++ src/taurus/dom/element.c
@@ -520,7 +520,10 @@ TaurusElement taurus_element_get_child(TaurusElement elem, uint16_t index) {
 
     /* Walk the child linked list */
     TaurusElement child = taurus_element_get_first_child(elem);
-    for (uint8_t i = 0; i < index && child; i++) {
+    /* size_t counter: a uint8_t wraps for index > 255 and the walk
+     * never terminates early as intended (CodeQL comparison-with-
+     * wider-type). */
+    for (size_t i = 0; i < index && child; i++) {
         child = taurus_element_get_next_sibling(child);
     }
